Extract module for namespace
